aibox-record/pyproject.toml
[tool.poetry]
name = "aibox-record"
version = "0.1.0"
description = ""
authors = ["zhaokefei <zhaokefei@tezign.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
loop-pycoral = {version = "^0.1.3", source = "looptech"}

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[[tool.poetry.source]]
name = "looptech"
url = "http://47.116.14.85:9000/simple"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
